OnLazyBackgroundPageActive must ensure the close_sequence_id is unique to avoid any outstanding OnLazyBackgroundPageIdle callback will not incorrectly send ExtensionMsg_Suspend.
Two errors are fixed in this patch:

1. A unique last_background_close_sequence_id_ is maintained
by ProcessManager. This corrects the error that previously a
BackgroundPageData struct was cleared and had its ID reset,
allowing it to then be incremented again to match the ID of
an outstanding callback.

2. OnLazyBackgroundPageActive previously checked for an
ExtensionHost object. However, when first launching an app
the methods are called OnLazyBackgroundPageActive then
OnBackgroundHostCreated.  Thus, the
last_background_close_sequence_id_ is now incremented
in OnLazyBackgroundPageActive without checking for an
ExtensionHost object.

BUG=316097
TEST=Manual test required as per https://code.google.com/p/chromium/issues/detail?id=316097#c10

Review URL: https://codereview.chromium.org/387863006

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@283053 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 77a538f344f505567429842fc0b943e67213da21

+++ extensions/renderer/dispatcher.cc
@@ -678,7 +678,7 @@ void Dispatcher::OnSetSystemFont(const std::string& font_family,
 }
 
 void Dispatcher::OnShouldSuspend(const std::string& extension_id,
-                                 int sequence_id) {
+                                 uint64 sequence_id) {
   RenderThread::Get()->Send(
       new ExtensionHostMsg_ShouldSuspendAck(extension_id, sequence_id));
 }
